Fix authentication (wow...)
