+++ app/root.tsx
@@ -1,23 +1,52 @@
-import type { LinksFunction } from "@remix-run/node";
+import type { LinksFunction, LoaderFunctionArgs } from "@remix-run/node";
 import {
   Links,
   Meta,
   Outlet,
   Scripts,
   ScrollRestoration,
+  json,
 } from "@remix-run/react";
-// import {
-//   darkColors,
-//   lightColors,
-//   theme,
-//   themeClass,
-// } from "./styles/theme/colors.css";
+
 import { ColorContext, useColorThemeState } from "./state/colorContext";
 import { Layout } from "./components/organism/Layout";
 
 import styles from "./tailwind.css?url";
 
 export const links: LinksFunction = () => [{ rel: "stylesheet", href: styles }];
+
+const isAuthorized = (request: Request) => {
+  const header = request.headers.get("Authorization");
+
+  if (!header) return false;
+
+  const base64 = header.replace("Basic ", "");
+  const [username, password] = Buffer.from(base64, "base64")
+    .toString()
+    .split(":");
+
+  return (
+    username === process.env.AUTH_USERNAME &&
+    password === process.env.AUTH_PASSWORD
+  );
+};
+
+export const headers = () => ({
+  "WWW-Authenticate": "Basic",
+});
+
+export const loader = async ({ request }: LoaderFunctionArgs) => {
+  if (!isAuthorized(request)) {
+    return json({ authorized: false }, { status: 401 });
+  }
+
+  // Load data for password-protected page here.
+
+  return json({
+    authorized: true,
+    // Include extra data for password-protected page here.
+  });
+};
 
 export default function App() {
   const contextState = useColorThemeState();
